+++ test/integration/shop/api/config_controller_test.rb
@@ -17,6 +17,8 @@ class Shop::Api::ConfigControllerTest < ActionDispatch::IntegrationTest
     assert_includes json.keys, "iframe"
     assert_includes json.keys, "integration_script_url"
     assert_includes json.keys, "operating_hours"
+    assert_includes json.keys, "tenant"
     assert json.dig("operating_hours", "is_open").in?([true, false])
+    assert json.dig("tenant", "display_address").present?
   end
 end